flask-app/requirements.txt
Flask==2.3.2
paho-mqtt==1.6.1
psycopg2-binary==2.9.9
Flask-SQLAlchemy==3.1.1